# refactor: change regexp simpler