- [x] check performance improvment

## before
```
2019-01-13 masking > 23GB_sqldump.sql  11376.61s user 153.19s system 98% cpu 3:14:28.25 total
```
## after
``` 
2019-01-13 masking > 23GB_sqldump.sql > 12529.03s user 280.19s system 99% cpu 3:34:28.94 total
```
 it indicates performance decrement. 🤔 

- [ ] check more detail of performance regression.